requirements.txt
Pillow==10.0.1
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
TgCrypto==1.2.5
